+++ cpp/CHANGELOG.md
@@ -10,6 +10,7 @@
 -   DTOs for Gmod, Codebooks, Locations, GmodVersioning, ISO19848
 -   `VisVersionsGenerator` tool for `VisVersions.h` and `VisVersionsExtensions.h` code generation
 -   `VIS` singleton class for accessing VIS versions and embedded resources
+-   `CodebookName` enumeration and conversion utilities for VIS codebook types
 
 ### Changed